Updated SEND, LIST and RECV for new architecture
